+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.1.1 (TBD)
 
-* TBD
+* Java 11
 
 ## [1.1.0](https://github.com/ngageoint/coordinate-reference-systems-java/releases/tag/1.1.0) (10-14-2021)